refactor: move CHAP.common.utils to CHAP.utils
